requirements.txt
flask>=3.0
psycopg2-binary>=2.9
bcrypt>=4.0
python-dotenv>=1.0
gunicorn>=21.0
